requirements.txt
apache-airflow==2.8.4
matplotlib
numpy
opencv_contrib_python==4.7.0.72
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
scikit_learn
seaborn
tensorflow==2.13.0
